requirements.txt
# Streamlit for the web app
streamlit==1.54.0

# Load environment variables from .env
python-dotenv==1.2.1

# OpenAI client for OpenRouter API
openai==2.24.0

# HTTP requests library (optional, useful for direct API calls)
requests==2.32.5